+++ covasim/people.py
@@ -4,7 +4,6 @@ Defines the Person class and functions associated with making people.
 
 #%% Imports
 import numpy as np
-import sciris as sc
 from . import utils as cvu
 from . import defaults as cvd
 from . import base as cvb
@@ -87,7 +86,7 @@ class People(cvb.BasePeople):
         # Initialize
         self.t = t
         counts = {key:0 for key in cvd.new_result_flows}
-        self.is_exp = cvu.true(self.exposed) # For storing the interim values since used in every subsequent calculation
+        self.is_exp = self.true('exposed') # For storing the interim values since used in every subsequent calculation
 
         # Perform updates
         counts['new_infectious']  += self.check_infectious() # For people who are exposed and not infectious, check if they begin being infectious
@@ -150,7 +149,7 @@ class People(cvb.BasePeople):
     #%% Methods for updating state
 
     def check_inds(self, current, date, filter_inds=None):
-        ''' Return indices for which the current state is false nad which meet the date criterion '''
+        ''' Return indices for which the current state is false and which meet the date criterion '''
         if filter_inds is None:
             filter_inds = self.is_exp
         not_current = cvu.ifalsei(current, filter_inds)
@@ -163,7 +162,7 @@ class People(cvb.BasePeople):
         ''' Check if they become infectious '''
         inds = self.check_inds(self.infectious, self.date_infectious)
         self.infectious[inds] = True
-        self.rel_trans[inds]  = 1.0 # TODO: make this dynamic
+        self.rel_trans[inds]  = 1.0
         return len(inds)
 
 
@@ -233,14 +232,16 @@ class People(cvb.BasePeople):
         ''' Check for whether someone has been contacted by a positive'''
 
         if self.pars['quar_period'] is not None:
+            not_diagnosed_inds = self.false('diagnosed')
+            all_inds = np.arange(len(self)) # Do dead people come out of quarantine?
 
-            # Perform quarantine
-            inds = self.check_inds(self.quarantined, self.date_known_contact) # Check who is quarantined
+            # Perform quarantine - on all who have a date_known_contact (Filter to those not already diagnosed?)
+            inds = self.check_inds(self.quarantined, self.date_known_contact, filter_inds=not_diagnosed_inds) # Check who is quarantined, not_diagnosed_inds?
             self.quarantine(inds) # Put people in quarantine
             self.date_known_contact[inds] = np.nan # Clear date
 
-            # Check for the end of quarantine
-            end_inds = self.check_inds(~self.quarantined, self.date_end_quarantine) # Note the double-negative here
+            # Check for the end of quarantine - on all who are quarantined
+            end_inds = self.check_inds(~self.quarantined, self.date_end_quarantine, filter_inds=all_inds) # Note the double-negative here
             self.quarantined[end_inds] = False # Release from quarantine
             self.date_end_quarantine[end_inds] = np.nan # Clear end quarantine time
 
@@ -394,26 +395,31 @@ class People(cvb.BasePeople):
             trace_time (dict): # days it'll take to trace people at each contact layer - should have the same keys as contacts
         '''
         # Figure out who has been contacted in the past
-        never_been_contacted = cvu.true(np.isnan(self.date_known_contact))  # Indices of people who've never been contacted
+        never_been_contacted = self.not_defined('date_known_contact')  # Indices of people who've never been contacted
 
         # Extract the indices of the people who'll be contacted
         traceable_layers = {k:v for k,v in trace_probs.items() if v != 0.} # Only trace if there's a non-zero tracing probability
-        for layer,this_trace_prob in traceable_layers.items():
-            this_trace_time = trace_time[layer]
+        for lkey,this_trace_prob in traceable_layers.items():
+            if self.pars['beta_layer'][lkey]: # Skip if beta is 0 for this layer
+                this_trace_time = trace_time[lkey]
 
-            p1inds = np.where(np.isin(np.array(self.contacts[layer]['p1']),inds))[0] # Get all the indices of the pairs that each person is in
-            p2inds = np.unique(np.array(self.contacts[layer]['p2'][p1inds])) # Find their pairing partner
-            contact_inds = cvu.binomial_filter(this_trace_prob, p2inds) # Filter the indices according to the probability of being able to trace this layer
-            self.known_contact[contact_inds] = True
+                nzinds = self.contacts[lkey]['beta'].nonzero()[0] # Find nonzero beta edges
+                p1inds = np.isin(self.contacts[lkey]['p1'], inds).nonzero()[0] # Get all the indices of the pairs that each person is in
+                nzp1inds = np.intersect1d(nzinds, p1inds)
+                p2inds = np.unique(self.contacts[lkey]['p2'][nzp1inds]) # Find their pairing partner
+                # Check not diagnosed!
+                contact_inds = cvu.binomial_filter(this_trace_prob, p2inds) # Filter the indices according to the probability of being able to trace this layer
+                self.known_contact[contact_inds] = True
 
-            # Set the date of contact, careful not to override what might be an earlier date. TODO: this could surely be one operation?
-            first_time_contacted_inds   = np.intersect1d(never_been_contacted, contact_inds) # indices of people getting contacted for the first time
-            contacted_before_inds       = np.setdiff1d(contact_inds, first_time_contacted_inds) # indices of people who've been contacted before
+                # Set the date of contact, careful not to override what might be an earlier date. TODO: this could surely be one operation?
+                first_time_contacted_inds   = np.intersect1d(never_been_contacted, contact_inds) # indices of people getting contacted for the first time
+                contacted_before_inds       = np.setdiff1d(contact_inds, first_time_contacted_inds) # indices of people who've been contacted before
 
-            if len(first_time_contacted_inds):
-                self.date_known_contact[first_time_contacted_inds]  = self.t + this_trace_time # Store when they were contacted
-            if len(contacted_before_inds):
-                self.date_known_contact[contacted_before_inds]  = np.minimum(self.date_known_contact[contacted_before_inds], self.t + this_trace_time)
+                if len(first_time_contacted_inds):
+                    self.date_known_contact[first_time_contacted_inds]  = self.t + this_trace_time # Store when they were contacted
+                if len(contacted_before_inds):
+                    self.date_known_contact[contacted_before_inds]  = np.minimum(self.date_known_contact[contacted_before_inds], self.t + this_trace_time)
+
 
         return